Move test into its own folder so we can just write that folder
